Buildout configurations for testing and development
They are based on Plone 4.2, and the development buildout includes
tools for localization.

It is assumend that the buildout is run with a Python interpreter that
is isolated (e.g.: a virtualenv) to minimize the chance of version
conflict as Python and Zope libraries are extensively used for system
and desktop tools nowadays.