Attempt filling empty slot after a subscription is moved
This completes d463610, where moving a subscription between timeslots
was implemented but this was missed.

When a user moves a subscription between timeslots, a spot is open in
the source timeslot and we need to check if there are users in the
wait list.

The event subscription is just changed to IObjectMovedEvent, as
IObjectRemovedEvent is a subclass of it, but that also means care need
to be taken to not run the handler function on the other sibling
event, IObjectAddedEvent.

As the IObjectMovedEvent's are propagated to sublocations, the handler
function also needs to check if the originating object is the same as
the one being called for, since nothing needs to be done for those
cases.